requirements.txt
absl-py==2.1.0
accelerate==0.30.1
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.2.0
arel==0.3.0
astunparse==1.6.3
async-asgi-testclient==1.4.11
async-timeout==4.0.3
attrs==23.2.0
av==12.0.0
beautifulsoup4==4.12.3
bidict==0.23.1
bitsandbytes==0.43.1
bleach==6.1.0
cacheout==0.14.1
cachetools==4.2.4
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cssutils==2.11.1
cycler==0.12.1
dataframe-image==0.2.3
dataset-tools==0.0.87
defusedxml==0.7.1
distinctipy==1.3.4
distro==1.9.0
dm-tree==0.1.8
dnspython==2.6.1
einops==0.8.0
email_validator==2.1.1
exceptiongroup==1.2.1
executing==2.0.1
fastapi==0.109.0
fastapi-cli==0.0.4
fastjsonschema==2.19.1
ffmpeg-python==0.2.0
ffmpy==0.3.2
filelock==3.14.0
flash-attn==2.5.9.post1
flatbuffers==24.3.25
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.6.0
future==1.0.0
gast==0.5.4
gdown==5.1.0
geojson==3.1.0
gitdb==4.0.11
GitPython==3.1.43
giturlparse==0.12.0
google-pasta==0.2.0
gradio==4.31.5
gradio_client==0.16.4
grpcio==1.63.0
h11==0.14.0
h5py==3.11.0
html2image==2.0.4.3
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.3
idna==3.7
imageio==2.34.1
imagesize==1.4.1
importlib_resources==6.4.0
imutils==0.5.4
inflect==7.2.1
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kagglehub==0.2.5
keras==3.3.3
keras-core==0.1.7
keras-nlp==0.12.1
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
lxml==5.2.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
memory-profiler==0.61.0
mistune==3.0.2
ml-dtypes==0.3.2
more-itertools==10.3.0
mpmath==1.3.0
multidict==6.0.5
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
networkx==3.3
ninja==1.11.1.1
numerize==0.12
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvcc-cu12==12.3.107
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
openai==1.34.0
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
orjson==3.10.3
packaging==24.1
pandas==1.5.2
pandocfilters==1.5.1
pillow==10.2.0
platformdirs==4.2.2
protobuf==3.20.3
psutil==5.9.8
PTable==0.9.2
pycocotools==2.0.7
pydantic==2.5.0
pydantic_core==2.14.1
pydicom==2.4.4
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.8.0
Pympler==1.0.1
pynrrd==0.4.3
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-json-logger==2.0.7
python-magic==0.4.27
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-toolbelt==0.10.1
rich==13.7.1
rpds-py==0.18.1
ruff==0.4.4
safetensors==0.4.3
scikit-image==0.23.2
scipy==1.13.0
semantic-version==2.10.0
sentencepiece==0.2.0
shapely==2.0.2
shellingham==1.5.4
SimpleITK==2.3.1
six==1.16.0
smmap==5.0.1
-e git+https://github.com/plasmapotatos/wildfire-smoke-detection@1050f9dab3ddfa6c1f3e4beb684d8dc705fc5f56#egg=smoke_detection
sniffio==1.3.1
soupsieve==2.5
starlette==0.35.1
stringcase==1.2.0
supervisely==6.73.104
sympy==1.12.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.37.0
tensorflow-text==2.16.1
termcolor==2.4.0
tifffile==2024.5.10
tiktoken==0.7.0
tinycss2==1.3.0
titlecase==2.4
tokenizers==0.19.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.3.1
torchaudio==2.3.0
torchvision==0.18.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.41.2
trimesh==3.23.5
triton==2.3.1
typeguard==4.3.0
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
varname==0.13.1
watchfiles==0.21.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.0.3
wrapt==1.16.0
xmltodict==0.13.0
yarl==1.9.4
